+++ .gitignore
@@ -59,3 +59,4 @@ typings/
 
 .vscode/
 .DS_Store
+.next/
