+++ config/index.js
@@ -1,9 +1,9 @@
 'use strict'
 
 module.exports = {
-    devHost: '127.0.0.1',
-    devPort: 3333,
     port: 3000,
+
+    nextDir: 'next',
 
     redisHost: '127.0.0.1',
     redisPort: 6379,
