+++ next/pages/index.jsx
@@ -1,0 +1,11 @@
+import React, { Component } from 'react'
+
+export default class index extends Component {
+    render() {
+        return (
+            <div>
+                test
+            </div>
+        )
+    }
+}
