--- routes/lib/index.js
@@ -1,5 +1,0 @@
-'use strict'
-
-Router.get('/', async (ctx) => {
-    ctx.body = 'hello'
-})
